% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chambr-package.R
\docType{package}
\name{chambr}
\alias{chambr}
\title{Functions to read and handle chamber data in WACL.}
\description{
Provides a key \code{read_chamber} function to read chamber data. Also
functions to visualise and check chamber data.
}
