% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chamber.R
\name{plot_chamber}
\alias{plot_chamber}
\title{Visualise Chamber Data}
\usage{
plot_chamber(.data, drop_constant = F)
}
\arguments{
\item{.data}{WACL chamber data read using read_chamber.}

\item{drop_constant}{Should constant variables be plotted? (T/F, defaults to F)}
}
\value{

}
\description{
To ensure that the chamber data has been read in correctly, this function
uses ggplot2 to visualise all of the numeric variables against time. There
is also an option to drop values that remained constant throughout.
}
