% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chamber.R
\name{read_chamber}
\alias{read_chamber}
\title{Read Data from the WACL Chamber}
\usage{
read_chamber(file)
}
\arguments{
\item{file}{The path to the file.}
}
\value{

}
\description{
Reads data from the WACL chamber into R, fixing column names and data
types. There will likely be parsing warnings - this is to be expected
and nothing to worry about. It is due to the files being flanked by
incomplete data with fewer columns.
}
